Added the proposed structure of the XDQ document
The Express Data Queuing (XDQ) technical report/design document will
describe the design of two different eBPF based scheduling frameworks to
the Linux kernel. The former is a Qdisc started by Cong Wang, and the
second is an extension to XDP.

I've added bullet points that describe the initial aim of each section.
Some bullet points are questions that I would like answers to before I
write those points.

The part of the eBPF section that I have written is quite dry.
Therefore, I will rewrite it to make it flow a little better in the
subsequent versions.