queue-exp: Implement STFQ using EiffelPifo
This adds an implementation for STFQ using the EiffelPifo primitive, to
show how that differs from the straight PIFO approach. It works by
re-computing the rank of the flow based on the next packet in the flow
after a dequeue. Neatly demonstrating how the two Eiffel additions to the
PIFO primitive are actually both required for the model to be able to
express the same queueing algorithms as the PIFO: Queueing per flow without
the ability to change rank at dequeue would not work.

Also an interesting property of the stable sort we use for the PIFO is that
if a flow dequeues a packet and then gets its rank updated so it's the same
as that of the next flow in the list, the same flow will get to dequeue
another packet until its rank increases to strictly more than the next
flow. With unit-size packets that means we'll get 122112 order instead of
121212 if enqueueing three packets from each flow using STFQ.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>