queue_exp: eiffel: remove items and re-add them on dequeue
Instead of just updating the flow rank and sort()'ing the PIFO on dequeue,
always remove the flow entirely and re-add it if needed. This makes sure
that if the re-computed rank is equal to that of the next flow, the two
flows will end up alternating, so we get the proper 121212 ordering for
unit-sized packets instead of the 122112 order noted in the previous commit.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>