queue-exp: Implement a hierarchical version of the Eiffel queue
This adds an version of the Eiffel queue that supports hierarchical
dequeue, by recursively enqueueing itself to its parent when getting a
packet enqueued. Supporting that requires a number of changes to the Eiffel
and PIFO primitives to support this the recursion.

The three-queue policy illustrated in the Eiffel paper is implemented using
this scheme (except for the weights and the shaping). This is a (very)
rough illustration of how the scheme would look like in a 'qdisc' setup
where the hierarchy is defined out of band and the BPF programs "only" set
packet ranks and classifies packets.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>