queue-exp: Add procedural version of hierarchical Eiffel
This adds another implementation of the Eiffel hierarchy from the paper,
this time implemented procedurally using only the PIFO queue primitive and
some state maps. This is meant to illustrate the case where BPF implements
the full enqueue and dequeue logic and the PIFO primitive is just made
available as a map type or via a helper (the other state can be kept in the
existing map types).

The behaviour is not quite correct, as it uses the flow primitive
differently than the hierarchical version. However, it is close enough to
get a sense of how this programming model would work, and it's not clear
that the flow primitive should exist at all (as it can just be created as
another PIFO).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>