Refactored framework and added a FIFO and SRPT
I changed and refactored the framework; for now, it breaks the current
scheduling algorithms. I have provided two new example schedulers, FIFO
and SRPT, to show how the new framework works.

I split the data structures and the scheduling algorithms to make it
more transparent which part is which. I also removed the packet
awareness from the queues to make them able to have references to
packets, flows, or queues without needing helpers. This change makes it
easier to create hierarchical schedulers and to track flows in
schedulers like SRPT.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>