Rewrote the introduction
The old introduction was outdated and a bit too wordy about what the
paper is ultimately about. I prioritized talking about XDP and how this
work is driven by adding scheduling to XDP using BPF. I want to add
strong examples of why XDP needs queuing and emphasize the prior work on
which we base our framework in the next iteration of this section. So
far, that would primarily be the Eiffel and PIFO papers.