Replaced the BPF packet queuing figure
The old BPF figure was incomplete and wrong. The new figure shows the
BPF queuing framework and without discussing the difference between XDQ
and the Qdisc. We will need to discuss whether it should have XDP and
Qdisc in this figure or make it more straightforward with more detailed
figures in the XDQ and Qdisc sections.